third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-selection-after-delete-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel correctly updates selection on node removal.


Running: testUserDelete
Selecting node...
The following node is now selected: <DIV class="child2">
Deleting selected node like it was a user action...
The following node is now selected: <DIV class="child3">
Deleting selected node like it was a user action again...
The following node is now selected: <DIV class="child1">
Deleting last child like it was a user action...
The following node is now selected: <DIV class="left">

Running: testExternalDelete
Selecting node...
The following node is now selected: <DIV class="child2">
Deleting selected node externally...
The following node is now selected: <DIV class="child3">
Deleting selected node externally again...
The following node is now selected: <DIV class="child1">
Deleting last child externally...
The following node is now selected: <DIV class="left">
